// src/app/api/ui/enums/action-type.enum.ts
export enum ActionTypeEnum {
  BuyFromSearch = 18,
  ViewOfferDetail = 20,
  CustomerAuthentication44 = 149,
  ParticipantAuthentication44 = 150,
  CustomerAuthentication223 = 151,
  ParticipantAuthentication223 = 152,
  CustomerAuthenticationMz = 153,
  ParticipantAuthenticationMz = 154,
  ManagerEnterAsEmploye = 155,
  ManagerAddedOrganization = 156,
  ManagerStartModerateOffers = 157,
  ManagerDoneModerateOffers = 158,
  MyCatalogAddPriceList = 159,
  AddAdditionalAgreement = 160,
  HistoryAddAsOrganization = 161,
  ManagerClickEnterAsEmploye = 162,
  OffersBatchUpdate = 163,
  OffersSearch = 164,
  OffersProlongation = 165,
  OffersExportToExcel = 166,
  OffersSendingToModeration = 167,
  OffersDisapproving = 168,
  OffersApproving = 169,
  OffersRemoving = 170,
  CartAddFromFavourites = 171,
  OrderCreationConfirmation = 172,
  AddToCartFromFullCart = 173,
  ViewMainMenuCustomItem = 174,
  ReportsMenuItemStatisticsClicked = 175,
  ReportsMenuItemRegisterClicked = 176,
  ReportsTradeRegistryTabClicked = 177,
  ReportsSupplierRegistryTabClicked = 178,
  ReportsCustomerRegistryTabClicked = 179,
  ReportsTradeRegistryGenerateButtonClicked = 180,
  ReportsSupplierRegistryGenerateButtonClicked = 181,
  ReportsCustomerRegistryGenerateButtonClicked = 182,
}